+++ bfabric/src/bfabric/bfabric.py
@@ -32,7 +32,7 @@ from bfabric.config import BfabricClientConfig
 from bfabric.config.bfabric_client_config import BfabricAPIEngineType
 from bfabric.engine.engine_suds import EngineSUDS
 from bfabric.engine.engine_zeep import EngineZeep
-from bfabric.rest.token_data import get_token_data
+from bfabric.rest.token_data import get_token_data, TokenData
 from bfabric.results.result_container import ResultContainer
 from bfabric.utils.cli_integration import DEFAULT_THEME, HostnameHighlighter
 from bfabric.utils.paginator import compute_requested_pages, BFABRIC_QUERY_LIMIT
@@ -88,8 +88,9 @@ class Bfabric:
         """
         config, auth_config = get_system_auth(config_env=config_env, config_path=config_path)
         auth_used: BfabricAuth | None = auth_config if auth == "config" else auth
-        if engine is not None:
-            config = config.copy_with(engine=engine)
+        # TODO https://github.com/fgcz/bfabricPy/issues/164
+        # if engine is not None:
+        #    config = config.copy_with(engine=engine)
         return cls(config, auth_used)
 
     @classmethod
@@ -99,21 +100,24 @@ class Bfabric:
         config_env: str | None = None,
         config_path: str | None = None,
         engine: BfabricAPIEngineType = BfabricAPIEngineType.SUDS,
-    ) -> Bfabric:
+    ) -> tuple[Bfabric, TokenData]:
         """Returns a new Bfabric instance, configured with the user configuration file and the provided token.
+
         Any authentication in the configuration file will be ignored, but it will be used to determine the correct
         B-Fabric instance.
         :param token: the token to use for authentication
         :param config_env: the config environment to use (if not specified, see `from_config`)
         :param config_path: the path to the config file (if not specified, see `from_config`)
         :param engine: the engine to use for the API.
+        :return: a tuple of the Bfabric instance and the token data
         """
         config, _ = get_system_auth(config_env=config_env, config_path=config_path)
-        if engine is not None:
-            config = config.copy_with(engine=engine)
+        # TODO https://github.com/fgcz/bfabricPy/issues/164
+        # if engine is not None:
+        #    config = config.copy_with(engine=engine)
         token_data = get_token_data(client_config=config, token=token)
         auth = BfabricAuth(login=token_data.user, password=token_data.user_ws_password)
-        return cls(config, auth)
+        return cls(config, auth), token_data
 
     @property
     def config(self) -> BfabricClientConfig: